Add more instructions
However, it requires to change the way we differentiate the labels and the
instructions. For any alphanumeric sequence, check it is in a hashtbl, and
return the corresponding type, like label, instructions, etc. If it is not an
instruction, check if it is a pseudo instruction. If it is not, check if it
matches the label pattern. Otherwise, raise an exception.